requirements.txt
pytest
appium-python-client
requests
pytest-rerunfailures